added polls and design

+++ models/task_status.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, String, BigInteger
 
 from .base import Base
 
@@ -7,7 +7,7 @@ class TaskStatus(Base):
     __tablename__ = "tasks_status"
 
     id = Column(Integer, primary_key=True)
-    user_id = Column(Integer)
+    user_id = Column(BigInteger)
     uid = Column(Integer)
     name = Column(String)
     status = Column(String)
